Introduce zoning specification .docx parser
Converting the .doc to a .pdf and using tesseract to parse the bylaws
works, but results in image files instead of text. This makes it
impossible to copy/paste the bylaws using normal methods.

Instead, convert the .doc to a .docx and parse the bylaws from the
text content of the file. The resulting text files are both smaller
than the old image files and simpler for the user to copy.